fix(router): nest nip05 settings under nostr settings
